+++ src/main/java/nextstep/ladder/BooleanGenerator.java
@@ -1,4 +1,4 @@
-package nextstep.laddar;
+package nextstep.ladder;
 @FunctionalInterface
 public interface BooleanGenerator {
     boolean generate();
