+++ src/main/java/nextstep/ladder/Ladder.java
@@ -1,4 +1,4 @@
-package nextstep.laddar;
+package nextstep.ladder;
 
 import java.util.ArrayList;
 import java.util.Collections;
