+++ src/main/java/nextstep/ladder/RandomGenerator.java
@@ -1,4 +1,4 @@
-package nextstep.laddar;
+package nextstep.ladder;
 
 import java.util.Random;
