+++ src/main/java/nextstep/ladder/Users.java
@@ -1,4 +1,4 @@
-package nextstep.laddar;
+package nextstep.ladder;
 
 import java.util.List;
