+++ src/test/java/nextstep/ladder/UserTest.java
@@ -1,9 +1,7 @@
-package nextstep.laddar;
+package nextstep.ladder;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 class UserTest {
